third_party/blink/web_tests/platform/generic/editing/selection/selectAllChildren-expected.txt

foobar
PASS s.anchorNode is span1.firstChild
PASS s.anchorOffset is 0
PASS s.focusNode is span1.firstChild
PASS s.focusOffset is 0
PASS window.getSelection().toString() is "foo"
PASS window.getSelection().toString() is "foobar"
PASS s.selectAllChildren(null) threw exception TypeError: Failed to execute 'selectAllChildren' on 'Selection': parameter 1 is not of type 'Node'..
PASS s.selectAllChildren(window) threw exception TypeError: Failed to execute 'selectAllChildren' on 'Selection': parameter 1 is not of type 'Node'..
PASS s.selectAllChildren("") threw exception TypeError: Failed to execute 'selectAllChildren' on 'Selection': parameter 1 is not of type 'Node'..
PASS successfullyParsed is true

TEST COMPLETE
